Moved App to components and added product counter to header
