+++ React/Cart/src/components/Header.jsx
@@ -2,7 +2,7 @@ const Header = () => {
     return (
         <nav id='navbar'>
             <a href="/">Produits</a>
-            <a href="/cart">Panier</a>
+            <a href="/cart">Panier(TODO : Afficher la taille du panier)</a>
         </nav>
     )
 };
