+++ React/Cart/src/main.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
 
-import App from './App.jsx';
-// import App from "./solution/App.jsx";
+import App from './components/App';
+// import App from "./solution/App";
 
 ReactDOM.createRoot(document.getElementById('root'))
   .render(<App />);
