+++ React/Cart/src/solution/Header.jsx
@@ -1,10 +1,13 @@
 import { Link } from "react-router-dom";
+import { useContext } from "react";
+import CartContext from "./CartContext";
 
 const Header = () => {
+    const { state } = useContext(CartContext);
     return (
         <nav id='navbar'>
             <Link to="/">Produits</Link>
-            <Link to="/cart">Panier</Link>
+            <Link to="/cart">Panier({state.products.length})</Link>
         </nav>
     )
 };
